% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpin.R
\name{draw_key_pin}
\alias{draw_key_pin}
\title{Title draw a pin on a map}
\usage{
draw_key_pin(data, params, size)
}
\arguments{
\item{data}{the data}

\item{params}{params}

\item{size}{the size of the pin}
}
\description{
Calls `mz_colourpin` to create a `rasterGrob`
}
