% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpin.R
\name{geom_pin}
\alias{geom_pin}
\title{Title}
\usage{
geom_pin(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
