% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpin.R
\name{mz_colourpin}
\alias{mz_colourpin}
\title{Title Create a colour pin}
\usage{
mz_colourpin(colour = "darkred", opacity = 50)
}
\arguments{
\item{colour}{the colour of the pin}

\item{opacity}{the opacity of the pin}
}
\value{
a pin
}
\description{
Title Create a colour pin
}
